requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-environ==0.12.0
django-render-partial==0.4
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
psycopg2-binary>=2.9
gunicorn
whitenoise
drf-nested-routers